+++ src/main/java/com/sprint/mission/discodeit/controller/UserController.java
@@ -3,12 +3,12 @@ package com.sprint.mission.discodeit.controller;
 import com.sprint.mission.discodeit.controller.api.UserApi;
 import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.service.UserStatusService;
-import com.sprint.mission.discodeit.service.dto.binarycontent.BinaryContentCreateRequest;
-import com.sprint.mission.discodeit.service.dto.user.UserCreateRequest;
-import com.sprint.mission.discodeit.service.dto.user.UserDto;
-import com.sprint.mission.discodeit.service.dto.user.UserUpdateRequest;
-import com.sprint.mission.discodeit.service.dto.user.userstatus.UserStatusUpdateDto;
-import com.sprint.mission.discodeit.service.dto.user.userstatus.UserStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.binarycontent.BinaryContentCreateRequest;
+import com.sprint.mission.discodeit.dto.service.user.UserCreateRequest;
+import com.sprint.mission.discodeit.dto.service.user.UserDto;
+import com.sprint.mission.discodeit.dto.service.user.UserUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.user.userstatus.UserStatusUpdateDto;
+import com.sprint.mission.discodeit.dto.service.user.userstatus.UserStatusUpdateRequest;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
