+++ src/main/java/com/sprint/mission/discodeit/controller/api/ChannelApi.java
@@ -1,9 +1,9 @@
 package com.sprint.mission.discodeit.controller.api;
 
-import com.sprint.mission.discodeit.service.dto.channel.ChannelDto;
-import com.sprint.mission.discodeit.service.dto.channel.ChannelUpdateRequest;
-import com.sprint.mission.discodeit.service.dto.channel.PrivateChannelRequest;
-import com.sprint.mission.discodeit.service.dto.channel.PublicChannelRequest;
+import com.sprint.mission.discodeit.dto.service.channel.ChannelDto;
+import com.sprint.mission.discodeit.dto.service.channel.ChannelUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.channel.PrivateChannelRequest;
+import com.sprint.mission.discodeit.dto.service.channel.PublicChannelRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
