+++ src/main/java/com/sprint/mission/discodeit/dto/controller/MessageDto.java
@@ -1,8 +1,8 @@
-package com.sprint.mission.discodeit.controller.dto;
+package com.sprint.mission.discodeit.dto.controller;
 
 import com.sprint.mission.discodeit.entity.Message;
-import com.sprint.mission.discodeit.service.dto.binarycontent.BinaryContentDto;
-import com.sprint.mission.discodeit.service.dto.user.UserDto;
+import com.sprint.mission.discodeit.dto.service.binarycontent.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.service.user.UserDto;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
