+++ src/main/java/com/sprint/mission/discodeit/dto/service/binarycontent/BinaryContentFileResponse.java
@@ -1,0 +1,8 @@
+package com.sprint.mission.discodeit.dto.service.binarycontent;
+
+public record BinaryContentFileResponse(
+    String contentType,
+    byte[] fileBytes
+) {
+
+}
