+++ src/main/java/com/sprint/mission/discodeit/dto/service/channel/ChannelDto.java
@@ -1,8 +1,8 @@
-package com.sprint.mission.discodeit.service.dto.channel;
+package com.sprint.mission.discodeit.dto.service.channel;
 
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ChannelType;
-import com.sprint.mission.discodeit.service.dto.user.UserDto;
+import com.sprint.mission.discodeit.dto.service.user.UserDto;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
