+++ src/main/java/com/sprint/mission/discodeit/dto/service/channel/ChannelUpdateRequest.java
@@ -1,0 +1,8 @@
+package com.sprint.mission.discodeit.dto.service.channel;
+
+public record ChannelUpdateRequest(
+    String newName,
+    String newDescription
+) {
+
+}
