+++ src/main/java/com/sprint/mission/discodeit/dto/service/message/MessageUpdateRequest.java
@@ -1,0 +1,7 @@
+package com.sprint.mission.discodeit.dto.service.message;
+
+public record MessageUpdateRequest(
+    String newContent
+) {
+
+}
