+++ src/main/java/com/sprint/mission/discodeit/dto/service/user/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.user;
+package com.sprint.mission.discodeit.dto.service.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
