+++ src/main/java/com/sprint/mission/discodeit/dto/service/user/userstatus/UserStatusUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.sprint.mission.discodeit.dto.service.user.userstatus;
+
+import java.time.Instant;
+
+public record UserStatusUpdateRequest(
+    Instant newLastActiveAt
+) {
+
+}
