+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.service.dto.user.LoginRequest;
-import com.sprint.mission.discodeit.service.dto.user.UserDto;
+import com.sprint.mission.discodeit.dto.service.user.LoginRequest;
+import com.sprint.mission.discodeit.dto.service.user.UserDto;
 
 
 public interface AuthService {
