+++ src/main/java/com/sprint/mission/discodeit/service/ReadStatusService.java
@@ -2,9 +2,9 @@ package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.entity.ReadStatus;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusCreateRequest;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusDto;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.readstatus.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.service.readstatus.ReadStatusDto;
+import com.sprint.mission.discodeit.dto.service.readstatus.ReadStatusUpdateRequest;
 import java.util.List;
 import java.util.UUID;
