+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -9,13 +9,13 @@ import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.ReadStatusService;
-import com.sprint.mission.discodeit.service.dto.binarycontent.BinaryContentDto;
-import com.sprint.mission.discodeit.service.dto.channel.ChannelDto;
-import com.sprint.mission.discodeit.service.dto.channel.ChannelUpdateRequest;
-import com.sprint.mission.discodeit.service.dto.channel.PrivateChannelRequest;
-import com.sprint.mission.discodeit.service.dto.channel.PublicChannelRequest;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusCreateRequest;
-import com.sprint.mission.discodeit.service.dto.user.UserDto;
+import com.sprint.mission.discodeit.dto.service.binarycontent.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.service.channel.ChannelDto;
+import com.sprint.mission.discodeit.dto.service.channel.ChannelUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.channel.PrivateChannelRequest;
+import com.sprint.mission.discodeit.dto.service.channel.PublicChannelRequest;
+import com.sprint.mission.discodeit.dto.service.readstatus.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.service.user.UserDto;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
