+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserStatusService.java
@@ -5,9 +5,9 @@ import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserStatusService;
-import com.sprint.mission.discodeit.service.dto.user.userstatus.UserStatusCreateRequest;
-import com.sprint.mission.discodeit.service.dto.user.userstatus.UserStatusUpdateDto;
-import com.sprint.mission.discodeit.service.dto.user.userstatus.UserStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto.service.user.userstatus.UserStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.service.user.userstatus.UserStatusUpdateDto;
+import com.sprint.mission.discodeit.dto.service.user.userstatus.UserStatusUpdateRequest;
 import java.time.Instant;
 import java.util.List;
 import java.util.NoSuchElementException;
