--- src/main/java/com/sprint/mission/discodeit/service/dto/binarycontent/BinaryContentFileResponse.java
@@ -1,7 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.binarycontent;
-
-public record BinaryContentFileResponse(
-        String contentType,
-        byte[] fileBytes
-) {
-}
