--- src/main/java/com/sprint/mission/discodeit/service/dto/channel/ChannelUpdateRequest.java
@@ -1,7 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.channel;
-
-public record ChannelUpdateRequest(
-        String newName,
-        String newDescription
-) {
-}
