--- src/main/java/com/sprint/mission/discodeit/service/dto/message/MessageUpdateRequest.java
@@ -1,6 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.message;
-
-public record MessageUpdateRequest(
-        String newContent
-) {
-}
