--- src/main/java/com/sprint/mission/discodeit/service/dto/user/userstatus/UserStatusUpdateRequest.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.user.userstatus;
-
-import java.time.Instant;
-
-public record UserStatusUpdateRequest(
-        Instant newLastActiveAt
-) {
-}
