+++ src/main/java/com/sprint/mission/discodeit/storage/BinaryContentStorage.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.storage;
 
-import com.sprint.mission.discodeit.service.dto.binarycontent.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.service.binarycontent.BinaryContentDto;
 import java.io.InputStream;
 import java.util.UUID;
 import org.springframework.core.io.Resource;
