+++ src/main/java/com/sprint/mission/discodeit/storage/local/LocalBinaryContentStorage.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.storage.local;
 
-import com.sprint.mission.discodeit.service.dto.binarycontent.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.service.binarycontent.BinaryContentDto;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
 import com.sprint.mission.discodeit.util.FileUtil;
 import jakarta.annotation.PostConstruct;
